requirements.txt
pyserial
clickhouse-connect
